add libxml2-dev package to workaround llvm build bug
LLVM is doing their build wrong and won't find the libxml2.a we
just built. I submitted https://reviews.llvm.org/D45106 but
until that is accepted we can add this workaround and make sure
that the version of libxml2 we download is the same as the one in
apk.